requirements.txt
python-telegram-bot==13.15
environs==11.0.0
redis==3.2.1
requests==2.32.3